refactor: strategySection
