+++ apps/journeys-admin/src/components/TemplateView/TemplateView.tsx
@@ -14,7 +14,7 @@ import { useJourneysQuery } from '@core/journeys/ui/useJourneysQuery'
 
 import { Role } from '../../../__generated__/globalTypes'
 import { useUserRoleQuery } from '../../libs/useUserRoleQuery'
-import { StrategySection } from '../StrategySection'
+import { StrategySection } from '../../../../../libs/journeys/ui/src/components/StrategySection'
 import { TemplateGalleryCarousel } from '../TemplateGallery/TemplateGalleryCarousel'
 import { TemplateGalleryCard } from '../TemplateGalleryCard'
