+++ libs/journeys/ui/src/components/StrategySection/StrategySection.spec.tsx
