+++ libs/journeys/ui/src/components/StrategySection/StrategySection.stories.tsx
