+++ libs/journeys/ui/src/components/StrategySection/StrategySection.tsx
